CSV Reader now covers async use cases; eliminating AsyncReader
